changed the core to use pydantic instead of json schema
