+++ notifiers/providers/join.py
@@ -2,10 +2,10 @@ import json
 
 import requests
 
-from ..core import Provider
-from ..core import ProviderResource
-from ..core import Response
 from ..exceptions import ResourceError
+from ..models.provider import Provider
+from ..models.provider import ProviderResource
+from ..models.response import Response
 from ..utils.schema.helpers import list_to_commas
 from ..utils.schema.helpers import one_or_more
