+++ notifiers/providers/mailgun.py
@@ -1,7 +1,7 @@
 import json
 
-from ..core import Provider
-from ..core import Response
+from ..models.provider import Provider
+from ..models.response import Response
 from ..utils import requests
 from ..utils.schema.helpers import one_or_more
