+++ notifiers/providers/popcornnotify.py
@@ -1,5 +1,5 @@
-from ..core import Provider
-from ..core import Response
+from ..models.provider import Provider
+from ..models.response import Response
 from ..utils import requests
 from ..utils.schema.helpers import list_to_commas
 from ..utils.schema.helpers import one_or_more
