+++ notifiers/providers/pushover.py
@@ -1,7 +1,7 @@
-from ..core import Provider
-from ..core import ProviderResource
-from ..core import Response
 from ..exceptions import ResourceError
+from ..models.provider import Provider
+from ..models.provider import ProviderResource
+from ..models.response import Response
 from ..utils import requests
 from ..utils.schema.helpers import list_to_commas
 from ..utils.schema.helpers import one_or_more
