+++ notifiers/providers/simplepush.py
@@ -1,5 +1,5 @@
-from ..core import Provider
-from ..core import Response
+from ..models.provider import Provider
+from ..models.response import Response
 from ..utils import requests
